added storageaccount secret for uploading, and separated upload page to /publish and /upload
