+++ sites/geohub/.env.example
@@ -3,6 +3,9 @@
 # adding to private variables here, you should also add it in $lib/variables/private.ts
 AZURE_STORAGE_ACCOUNT=
 AZURE_STORAGE_ACCESS_KEY=
+# Storage account for data uploading
+AZURE_STORAGE_ACCOUNT_UPLOAD=
+AZURE_STORAGE_ACCESS_KEY_UPLOAD=
 DATABASE_CONNECTION=
 # Private dynamic variables for Azure Active Directory (imported via $env/dynamic/private in Auth.js)
 # these dynamic variables are loaded from `.env` in local, and they are loaded from application settings in server
