+++ sites/geohub/src/lib/server/helpers/generateHashKey.ts
@@ -1,0 +1,5 @@
+import crypto from 'crypto'
+
+export const generateHashKey = (data: string) => {
+  return crypto.createHash('md5').update(data).digest('hex')
+}
