+++ sites/geohub/src/lib/server/helpers/index.ts
@@ -9,3 +9,4 @@ export * from './pageNumber'
 export * from './getStyleById'
 export * from './getMapStats'
 export * from './getStarCount'
+export * from './generateHashKey'
