+++ sites/geohub/src/lib/server/variables/private.ts
@@ -1,3 +1,15 @@
-import { AZURE_STORAGE_ACCOUNT, AZURE_STORAGE_ACCESS_KEY, DATABASE_CONNECTION } from '$env/static/private'
+import {
+  AZURE_STORAGE_ACCOUNT,
+  AZURE_STORAGE_ACCESS_KEY,
+  AZURE_STORAGE_ACCOUNT_UPLOAD,
+  AZURE_STORAGE_ACCESS_KEY_UPLOAD,
+  DATABASE_CONNECTION,
+} from '$env/static/private'
 
-export { AZURE_STORAGE_ACCOUNT, AZURE_STORAGE_ACCESS_KEY, DATABASE_CONNECTION }
+export {
+  AZURE_STORAGE_ACCOUNT,
+  AZURE_STORAGE_ACCESS_KEY,
+  AZURE_STORAGE_ACCOUNT_UPLOAD,
+  AZURE_STORAGE_ACCESS_KEY_UPLOAD,
+  DATABASE_CONNECTION,
+}
