+++ sites/geohub/src/routes/data/publish/+page.ts
@@ -1,0 +1,2 @@
+export const csr = true
+export const ssr = false
